package.json: Include opencv.js
There's no official NPM repository for opencv, only option is the
pre-built verison as part of the docs. Add a fetch step to grab that.

It's huge, so load async. Any components will have to wait for it to be
ready.

We may want to compile it ourselves in the future, but that can happen
when it's needed.