client/ingest_form: Mark all individuals in one sample
Selecting samples one-by-one is a fraught with peril in retrospect.

* What was the last individual again?
* Wait, did I select the wrong one?
* Help! I want to go back.

Instead, add a bounding box for every individual, labelled with the
individual number, and use these selections to replace the drop-down
box.

Reconstruct the metadata_api to just have one call, that fetches all
data for a sample.

Remove the advancing logic, and "save and next" button is just save
now.