search: Generate search form to control querystring #28
As part of the initial page-fetch, get some information on available
metadata, hand that clientside where we can combine with the current
querystring and metadata_api to make a search form.

The server could nearly generate the form itself, but wouldn't have the
translations for the metadata fields (at least not currently), so we#d
have to do some replacement anyway.