client/annotate_selectbar: Add selectbar to annotate view #37
Hand the filter querystring from the search page to the annotation
page.

Fetch search results again, this time populating a dropdown with all
results, each of which is a link to a different individual.

This isn't particuarly efficient, as we're continually re-fetching the
search results. But it at least demonstrates what I think we'd like.