client/ingest: Simplify bounding_box labelling
Instead of handing the full individual to the viewer, so it can decide
on a title, move this decision back to ingest_form, which then asks the
metadata_api what the title should be.

Simplify the field names, as we don't need the pseudo-JSON, which also
means that we can use the label as part of the field identifier (since
we only need to parse the prefix).

As the indentifier is part of the field name, we can parse it out
rather than needing the full list of individuals when rebuilding
bounding_box lists, so the load_individuals event just tells the viewer
to rebuild it's list of objects.