client/viewer/editable_line: Angle-based snapLine replacement #31
The lerp() based betweenPoints was struggling. Being distance based it
wasn't really possible to work out the ordering of points.

Switch to snapLine, which works out the difference in angle between a
point on the expected line and the given point. In the process:

* End points will use the previous 2 points to get their angle, and
extrapolate from there
* Use the angle to find out if we're now "behind" the origin point, and
don't go any further.