docker: Ignore all gitignores #5
Symlink .gitignore -> .dockerignore, so the COPY command will ignore
~anything non-git.

The main reason for doing this is to not blat the .venv created in the
image with a local version.

This means that ".dir" directories won't be available, so recreate them
after COPY.

In the process tidy up the volumes, db should have always been a wsgi
mount, not nginx.