annotate: Add authority to annotations, users #30
Add a numeric authority field to annotations, to indicate if this was
done by an expert / non-expert, human / non-human user. Set this on
annotation save.

Add a "species_expert" option to a user's profile, choosing from any
one of the available species Taxonomy entries. Use this to decide what
AnnotationLevel to select when saving an annotation.

Obviously this hard-codes the existence of a species taxonomy, but not
having one is incredibly unlikely. It would still be possible to use
the site without anyway, but the expert option would be non-functional.