search: Remove open check when getting search results #61
We need to fetch the search results when showing the annotation page,
so we can't hide them at this point.

This check should eventually be redundant anyway, even if you list the
individuals that doesn't give you access to the annotations within. And
these should be further hidden by other tickets.