ingest: Allow ingest of single indivduals #63
An image might be of a single individual, in which case you could drag
away all bounding boxes and put the one you want back, but that would
get laborious.

Instead, offer a dropdown to choose a single individual to be visible,
when selected we only ingest that one, rather than the whole lot.