Move label generation server-side #63 #73
Not being able to generate a textual string server-side is problematic
in the admin page & when generating a list of choices for existing
projects.

Given we've enshrined the notion of slide / individual in the UI, make
individualLabel also a expected field, and use it to generate labels
server-side.

Remove individualTitle from the metadata_api, so we only do this in one
place.