client/meta: Format dates with django locale #86
Browser locale was a poor assumtion, as desktops in Hafro are
configured with en-US by default, even though this results in
incorrectly formatted dates for the users.

Instead, switch our version of english to be explicitly "en-GB" in
Django. Strip off "-GB" when just used as translations, and feed the
full string into dt locale libraries.