client/viewer: Translate incoming image to blob
The image download link & image upload API currently try to send the
original blob, which for a video stream is the <video> element. The
download link special-cases this, the upload doesn't.

Instead, convert anything that wasn't Blobby into a Blob on load(), and
use this blob in the ingest form & download link.

This also gives us an opportunity to recreate a blob if we want all
images to be in JPEG format in the future, e.g.