client: Make ingest metadata editable #89
Rework meta.js so we create input fields rather than fixed strings,
and turn this on for the ingest metadata table.

Instead of directly submitting these values to server-side, manually
sync changes between fields & the hidden JSON blobs, and submit those.

Eventually we may want to get rid of this and submit more normally, but
do that in it's own step.