docker: Destroy & rebuild staticfiles on startup
When starting a new image, the static volume may already exist. If so
we need to repopulate it with latest content.

To be able to do this, the app user needs to be able to write to the
volume, chown the default volume content to app, so subsequent
collectstatic runs can work.