docker: Make static an anonymous volume
Anonymous volumes are more ephemeral, and recreating on rebuild is more
accepted. Use volumes_from: to expose this, and the other volumes we
want anyway, to NGINX.

This in theory would be enough to not need the previous
recreate-at-runtime commit, but docker compose will attempt to recycle
previous anonymous volumes. So keep it anyway.