client/ingest: Add auto-crop #112
Implement a cropping algorithm similar to a flood-fill that expands a
bounding box until we find a box where nothing matches the reference
pixel.

Hard code tolerances for now and see if this is sufficient. Both
tolerance & rescale could be configurable sliders in future, but see if
these guesses cover most of the cases first.